Add files via upload
The driver abnormality monitoring system developed is capable of detecting drowsiness, drunken and reckless behaviors of driver in a short time. 
The Drowsiness Detection System developed based on eye closure of the driver can differentiate normal eye blink and drowsiness and detect the drowsiness while driving. 
The proposed system can prevent the accidents due to the sleepiness while driving. 
Information about the head and eyes position is obtained through various self-developed image processing algorithms. 
During the monitoring, the system is able to decide if the eyes are opened or closed. 
When the eyes have been closed for too long, a warning signal is issued. Processing judges the driver’s alertness level on the basis of continuous eye closures.